rename: useCalculateFirstViewAnimatedSize.ts ->  useAddFirstViewAnimation.ts #12
