+++ src/hooks/useAddFirstViewAnimation.ts
